sepolicy_vndr: add policies for Spearhead and DLSC
Spearhead framework is used for collecting data path stats
for debugging. DLSC is the agent that will collect stats
related to the network stack. For this framework to function
we need to define new SELinux policies to provide the
required access privileges.

Since Spearhead will only be enabled on userdebug
builds, move the Spearhead framework policy to
the test folder.

Change-Id: Icb29b7ffd73c255933d05fe7e5de31aa724e2b73